requirements.txt
langchain
langchain-community
streamlit
chromadb
pypdf
langchain-text-splitters
sentence-transformers
vaderSentiment
langchain-chroma
langchain-huggingface
langchain-ollama
pynput
Pillow
pytesseract
tesseract-ocr
Pyside6
SpeechRecognition 
PyAudio 
pyttsx3 
openai-whisper